+++ NEWS.md
@@ -5,6 +5,7 @@ Red Panda C++ Version 0.14.0
   - enhancement: greatly speed up code completion 
   - fix: code folding calcuation not correct when some codes are folded and editing after them
   - enhancement: code completion ui redesigned
+  - fix: mainwindow action's short cut doesn't work,  if the action is not in menu or toolbar
 
 Red Panda C++ Version 0.13.4
   - fix: when copy comments, don't auto indent